Updated openstack/openstack Project: openstack/python-barbicanclient  424f2ac07136d9e50c377db52f4a7e1f0c3d9191
Fix keystone version selection

The client currently only accepts '3.0' which directly conflicts with
other clients (like OSC). It is not currently possible to export
OS_IDENTITY_API_VERSION and have it work properly for both clients.

In addition, the way it is written now "DEFAULT" is not really a
proper default, as 2.0 is the fallback. This change resolves that
inconsistency.

Closes-bug: #1496597
Change-Id: I0d829e46fc251eb7762912c7f89f79cb267e6abe