Updated openstack/openstack Project: openstack-infra/shade  73a83fdcff98f7f6949e9d8e553ea1244906428c
Fix exception lists in functional tests

There is a pattern in our functional tests where exceptions are
aggregated within a list and then joined within a single string.
However, the join expects them as strings, not as exception objects.
When this code is hit, it causes an exception:

TypeError: sequence item 0: expected string, OpenStackCloudException found

This change converts those exception objects to strings.

Change-Id: I0a041a61537b79c73615e5136492141da0926fee