Updated openstack/openstack Project: openstack/barbican  b25421e0aad66696c20c0a7070a39fda244bdc85
Integrated with PyKMIP Pie API

The PyKMIP Pie API provides a simpler interface for communicating with
KMIP devices. The interface is simplified to allow KMIP integrators to
quickly be able to communicate with a KMIP device. It does not provide
all of the KMIP optional parameters, but those features are not
required for integration with KMIP in the KMIP secret store.

Depends-On: I568a28288068c6016dc67244d979d3619874a425
Change-Id: Ie6f71752eb4d7c7f01942b3881db7a4aa424cb49