Updated openstack/openstack Project: openstack/manila  a77d15222c5c801bdba3e24562fe0721798e0c63
Add possibility to setup password for generic driver

Make it possible to setup password for generic driver service image and
scenario tests. Also allow to just disable its usage not unsetting the password

It is useful for testing purpose when we have some problems with SSH keys
injection to Nova VMs.

Change-Id: I26dd32b1de8cceeaa6dc674092efec683df71889
Related-Bug: #1480870