Updated openstack/openstack Project: openstack/ceilometer  4fcb80c8ee7de2e820374ac4e11a1d93305b988b
Change tox default targets for quick use

This change makes it so that if you call `tox` or `tox -r` you will
get a useful result as a developer running tests:

    * py27 and py34 unit tests
    * a single functional test run (using mongodb)
    * pep8

This is slightly different from what was there before which was
as above but _two_ functional test runs, with mysql and postgresql.

The goal here is to make it dead simple for someone to (as quickly
as possible) make sure that there code is okay before committing.

Change-Id: I27dcf5265dea18be541b8b2857b0a135ff22f499