Updated openstack/openstack Project: openstack/horizon  606a443f45d5193fe935dd31062dc11ed19bb7da
Improve message propagation in exceptions.handle

In horizon/exceptions.py the handle(message="blah") function only
passes the message onto the user if the exception type is
HorizonException or a specifically handled class. In many cases,
including many forms, the except block calling this method is
catching the broad Exception and not more specific errors. Once
Exception is caught, we pass on an error message to
horizon/exceptions.handle(). That message is ignored in several cases
and the result is the underlying Django forms puts up a generic error
message instead: "Danger: There was an error submitting the form..." and
the redirect still happens.

This change treats exceptions where a message is passed in to handle as
recoverable as the exception has been caught and a user intended message
attached.

Closes-Bug: #1484723
Change-Id: I763f7606dde5df6cf09cbaf2c8e90158e0f256e3