Updated openstack/openstack Project: openstack/nova  a988f844baeff1cb6188fd5277a2d594551a320a
Fix an issue with NovaObjectRegistry hook

Due to the order in which the registration hook is called from
oslo.versionedobjects' registration handler, our existing method
does not always make the proper decision when multiple versions of
the same object are registered. This is actually a bug in o.vo,
but we can rewrite our method to be more deterministic such that
it works now, and after we fix the ordering thing in o.vo.

This patch makes it set objects.OBJECT if the new class being
registered is newer, instead of being based on the order of the
classes in the registry at the time the hook is called.

Change-Id: Ic1f9bf5115d72244e317c8fe908aa720e7f9a1a3