Updated openstack/openstack Project: openstack/nova  69bcfc3d3e679fc2ee4e320f6efedb747af2536b
Add Instance and InstanceList v2.0 objects

This adds Instance and InstanceList v2.0 and moves v1 compat-related
tests into a separate test subclass.

Note that since services like api and conductor will be sending v2
objects to older computes, this also adds a backport-to-1.x case
to each object for compatibility with kilo computes. Hopefully in the
future we'll be able to avoid needing to do that by capping object
verisons according to the service version. For now, the structure
of instance hasn't diverged enough so this easy approach works.

Related to blueprint liberty-bump-object-and-rpcapi-versions

Change-Id: Icfd2529962c0430761d2424069f0fd60a1b5c260