Updated openstack/openstack Project: openstack/neutron  84497a65f3886083d08a3dfecee7f7b5f4c655c1
_bind_devices query only existing ports

If a port is deleted right before _bind_devices is called,
get_ports_attributes will throw an exception since the row
corresponding to the port doesn't exist in the OVS DB.
Avoid that setting if_exists to True. The port will be
processed as deleted by the agent in the following iteration.

Change-Id: Ia6590d76f8683e6cba562cde3c39b051549f6c04
Closes-bug: #1489014