Updated openstack/openstack Project: openstack/cinder  31ff5b05879fcec5124291816077660aa46f30f5
Error message in update_migrated_volume was incorrect

If an error was encountered in self.vg.rename_volume the error
message tried to use volume['name'], which is an invalid key.
The invalid key error would then cause the retype/migration to
abort in the middle of the process and leave original and temporary
volume in place. This patch fixes this bug by providing a valid key.

Change-Id: I73e1335c649df41d5167efc3cb4b9af48a5d537d
Closes-bug: #1492462