Updated openstack/openstack Project: openstack/heat  49c4d39516e322a111764234d06cdc88ddd7c804
Move Resource exceptions to common module (1)

It is convenient to have all exceptions in exception module.
Also it is reduces namespace cluttering of resource module and decreases
the number of dependencies in other modules (we do not need to import resource
in some cases for now).
UpdateReplace exception is moved in this patch.

Change-Id: Ief441ca2022a0d50e88d709d1a062631479715b7