Updated openstack/openstack Project: openstack-dev/devstack  42bff13ecfe16f1f8e9d13e0c81ff34174e584a3
Disable the shelve tests for nova-cells using tempest config

Change from using a blacklist to disable the shelve tests for
nova-cells to using the tempest config option to disable the shelve
feature tests.  This is the intended method of disabling feature tests.
This first commit is to add code to lib/tempest to disable the shelve
feature test if the nova-cells service is enabled.  The next
st will remove the shelve blacklist from
http://git.openstack.org/cgit/openstack/nova/tree/devstack/tempest-dsvm-cells-rc

Change-Id: Ibf1f9aaa63e5f17b7d8774b511940ba8421e0887
Partial-bug: 1491152