Updated openstack/openstack Project: openstack/heat  88c8cbe03b8703c4a4d242259a0329e84b01b169
Move Resource exceptions to common module (2)

It is convenient to have all exceptions in exception module.
Also it is reduces namespace cluttering of resource module and decreases
the number of dependencies in other modules (we do not need to import resource
in some cases for now).
ResourceUnknownStatus exception is moved in this patch.

Change-Id: Iad44c16252f5171984859f0ddd05b821f5dfc7e7