Updated openstack/openstack Project: openstack/heat  03c76a1f474b78855bac235a3c4121e360a2a931
Move Resource exceptions to common module (3)

It is convenient to have all exceptions in exception module.
Also it is reduces namespace cluttering of resource module and decreases
the number of dependencies in other modules (we do not need to import resource
in some cases for now).
ResourceInError exception is moved in this patch.

Change-Id: I122c032468efeb97d165af4ae372bc35b60a11a2