Updated openstack/openstack Project: openstack/heat  32026ab3eeca9710db8e76f76313a94a75578222
Move Resource exceptions to common module (4)

It is convenient to have all exceptions in exception module.
Also it is reduces namespace cluttering of resource module and decreases
the number of dependencies in other modules (we do not need to import resource
in some cases for now).
UpdateInProgress exception is moved in this patch.

Change-Id: If694c264639bbce5334e1e6e7403b225ce1d3aee