Updated openstack/openstack Project: openstack/heat  68fc18b62b6534bcbbbb85b5dc93e91399064720
Move Resource exceptions to common module (5)

It is convenient to have all exceptions in exception module.
Also it is reduces namespace cluttering of resource module and decreases
the number of dependencies in other modules (we do not need to import resource
in some cases for now).
NoActionRequired exception is moved in this patch.

Change-Id: Idf9bfaa8a51dd6ea679a0148c712529e3c3b072a