Updated openstack/openstack Project: openstack/cinder  1e318b03e10704cd6a60243ba4a292312fb228c9
Switch SVC driver to use lsportfc to determine FC target WWPNS

Switched the startup code within SVC driver to use the
SVC lsportfc command rather than the lsnode command to determine the
target WWPNs for an SVC node, and updated SVC testcases.
The SVC lsportfc command returns all of the FC ports so that
all of the FC ports can be zoned and thus utilized by
the VM that a volume is attached to.
Where the lsnode command is documented to only report the
first 4 FC ports.

Closes-Bug: 1490041

Change-Id: I6acac0080a96a5b3dfbc6fb8b54a456206b21688