Updated openstack/openstack Project: openstack-infra/project-config  107c6a7a4f8c2451d0ad405e63e015b61c3c541e
Remove ansible-lint-jobs from infra-ansible gate and enable experimental job

Currently ansible-lint-jobs does not work with
the infra-ansible repo, that job is only meant
for ansible roles.

This patch removes ansible-lint-jobs from the check
and gate queue for infra-ansible and set noops to
allow merges.
It also introduces the gate-infra-ansible job on
the experimental queue for testing, would enable
that on check/gate when we get confidence on it.

Change-Id: I8b6871d24e79817685e1ca2b21d4334ac9ad5e39