Updated openstack/openstack Project: openstack/security-doc  25b53c1b2a39855a59c2e4ddbd795970514f37a5
Correct the keyword defences to defenses

The keyword defences is inadvisable.
It's better to be modified as defenses.
Even though the meaning of defences is same as defenses,
but if be defenses the whole terms in this section are
unified as defense or defenses. And it can coincide
with the section title "Layered defenses" more.

Change-Id: Id3a0b27dd5ce1c789638f83eebba76f62eefcede
Closes-Bug: #1491036