Updated openstack/openstack Project: openstack/heat  206786a0ed66b053003a3a63edb3c0d0a262a595
Start scaling down in an InstanceGroup rolling update ASAP

When the effective batch size is reduced due to a dwindling number of
resources we need to update, we may be able to reduce the size of the group
prior to the final batch without violating the minimum number of members we
are required to keep in service. This change ensures that any excess nodes
are removed as soon as possible.

Change-Id: Ic7b9544825a92b8720dcff15fe581038dff43582
Partially-Implements: blueprint scaling-group-common