Updated openstack/openstack Project: openstack/cinder  8e70d0d286cca00f770bdbb8c2fd61ba32fd1c74
Add os-brick's scsi_id command to Cinder rootwrap

FC drivers are failing on calls to scsi_id. This command
is in the os-brick rootwrap filters, but that file is not
being pulled in to Cinder yet.

As a temporary fix until that work can be completed, this
patch adds the missing command directly to volume.filters
to allow this command to run. Follow up work will be needed
to remove this and other legacy brick commands from the
volume.filters file and instead pull in os-brick's
os-brick.filters values.

Change-Id: Iba8e5c82f8d203759cda0e21de5fdf2404b18e69
Partial-bug: #1479842