Updated openstack/openstack Project: openstack/heat  846505eae0e68ee04155c4584d204927a6b1fc83
Don't do unnecessary updates in ResourceGroup rolling updates

During a rolling update of a ResourceGroup, there may be extra resources
around (due to a simultaneous scale-down) that contribute to our
min_in_service but which are fated to be deleted. There is no point
updating these. This patch changes to algorithm to only update resources we
actually care about. (Happily, this makes the logic a lot simpler too.)

Change-Id: I4eedccfb0ca3b6d6eab3a777635836c39cda76cd
Partially-Implements: blueprint scaling-group-common
Closes-Bug: #1489192