Updated openstack/openstack Project: openstack/heat  fd45822c09ddebef3420f3d46a0a9454ea0b7c17
Merge Server os-collect-config metadata on create

Currently if someone does this, we overwrite the template metadata:

  aserver:
    type: OS::Nova::Server
    metadata:
      os-collect-config:
        polling_interval: 10

This technique may be useful to provide additional configuration data
to os-collect-config, other than the credentials etc heat provides
internally, so instead of overwriting the os-collect-config map
merge with anything defined in the template (if it exists).

Note if there are collisions between the template and the keys used
internally the internal data will take precedence (same as current
behavior but without overriting the entire map).

Change-Id: I2480e716d2685bfb2376369d525f681fd6557e2e
Partial-Bug: #1488596