Updated openstack/openstack Project: openstack/python-ironicclient  c9dd5e3123adb00d6275db50f7bb5a10b5f9b319
Fixes file cache TypeError

The internal cache (common.filecache) uses dogpile.cache's dbm backend,
which uses the bsddb package from the standardlib (in py2 only). bsddb
however breaks with a TypeError if the cache key is a unicode string.

This change fixes this by using the ``key_mangler`` arg on cache
creation, which (according to the dogpile.cache docs) is to be used with
dbm/bsddb.

Change-Id: Ie4911e31567d773357da99af3ded0548ad0b7db1
Closes-Bug: #1492205