Updated openstack/openstack Project: openstack/neutron-vpnaas  2baf733818fcaa103e497cb659dd130dbe862569
Manage cleanup of .ctl/.pid files for LibreSwan

LibreSwan checks for the presence of pid/ctl files when starting up and
will error out if they already exist. However, LibreSwan's usage of the
capabilities library removes the access required to cleanup .ctl and
.pid files on shutdown if any of the directories in the path are missing
explicit permissions for root. This is not considered a bug by the
LibreSwan maintainers, so the LibreSwan driver must work around it by
checking if it is okay to remove the files and removing them on
startup. It must also wait for shutdown to complete before restarting
the daemon on LibreSwanProcess.restart().

Introduces new configuration for retrying check for process shutdown.

DocImpact


Change-Id: I5c215d70c348524979b740f882029f74e400e6d7
Closes-Bug: #1331502