Updated openstack/openstack Project: openstack/cinder  d18c82e8effbd8200474c00289fffa751adc3e9c
Add mechanism to update snapshot provider_id

During the L release we added the ability to store
provider_id info for volumes and snapshots.  We did
not however add a mechanism for this info to be updated
on upgrades.

This patch adds on to the volume update mechanism and
gives a driver the opportunity to set provider_id columns
that aren't set on snapshots just as it does for volumes.

Change-Id: Ie6585e4f0d2a4a511e5fbe35b91186d8ad7e9874
Closes-Bug: #1489182