Updated openstack/openstack Project: openstack-infra/subunit2sql  a9c22f0396b2b74fbf69dfc70b5db91ae432a423
Add option to subunit2sql for attr regex

This commit adds a new CLI flag/config option to specify a regex for
the attributes in the test_id. This mechanism works by doing a regex
match on the test id, it takes the first match group as the contents
of attrs, and removes the entire match from the test_id stored in the
db. Previously subunit2sql was hardcoded to use the default from the
python subunit.run stack which is to have the attrs inside of '[ ]'
in the test_id, however nothing in the subunit protocol dictates that
attrs are specified in this manner. So to accomodate other test run
environments this makes that aspect of subunit2sql configurable.

Change-Id: Ia3e41ad20ad1c105b8dd2f345f8ed45e3de8186c