Updated openstack/openstack Project: openstack/nova  cd65b5fd55dc25c44dc9f8ad7e31c6edf1610dbd
libvirt: cleanup() serial_consoles after instance failure

When nova cleanup() is called to release all resources related to an instance
it can thow InstanceNotFound.  Which is not caught and therefore returned to
the user.

This is because the generator function (_get_serial_ports_from_instance)
is not *executed* under try/except.  While that is clearly the intent of the
code.

Change-Id: I9e9410ab7eec5fa667c0c3de4548c49df050c167
Closes-Bug: 1480514