Updated openstack/openstack Project: openstack/nova  f560f3bcdc756c1ef8fe5d1c3011a5c34f538531
libvirt: Fix snapshot delete for network disk type for blockRebase op

_volume_snapshot_delete was raising an exception as part of blockRebase
op when the new backing file happens to be None, which is a valid
scenario.

This patch ensures that we don't look for the libvirt disk spec (skip
calling _get_snap_dev()) if the new backing file for rebase op is
None.

Change-Id: I98deda75310f0b44b70257071d282aa50babe06b
Tested-By: Zhang Jinnan<zhang.jinnan@99cloud.net>
Closes-Bug: #1477110