Updated openstack/openstack Project: openstack/ironic-python-agent  46eb0dfa354e210cff174901fe909fb339cafebf
Tuples returned by clean steps cause API error

If a clean step returns a tuple (like the output of
processutils.execute), WSME raises an error about the type. Casting
the result as a list seems to make it work. This is almost certainly
a bug in WSME, but after digging a bit I couldn't find one.

Change-Id: I91cb7a989ab56c0942469e03937945d4a590ad95
Partial-bug: #1493554