Updated openstack/openstack Project: openstack/trove  7870e1d222323d6502eb2c69d36ae45da34c42e3
Update ignore_dbs for MySQL 5.6

'lost+found' directories are no longer needed in the ignored list.
Include 'performance_schema' system database (by default ON in >=5.6.6).
Set 'performance_schema = ON' in default Trove configuration template and
add a validation rule for it so that users can turn it OFF if they so
desire.

Change-Id: Iecc114594091eabf8a0664e066b9de22f95d1e81
Closes-Bug: 1486721
Related-Bug: 1370646