Updated openstack/openstack Project: openstack/requirements  844140fedc99d00a04ddfdfdaa49088aecba939a
Bump the version of python-muranoclient to 0.7.0

A new release of muranoclient is out, having several new features
implemented. Other projects (such as murano-dashboard and murano
itslef) have to bump the minimum version of muranoclient to utilize
these features.
This patch bumps the minimum requirement and pins upper contraint to
the same value.

Change-Id: I270a9d8c321556db6da747e652d034b14390b83f