Updated openstack/openstack Project: openstack/keystone  06730f12f3b2f160552806cdd9d237f5489553a8
Remove padding from Fernet tokens

Fernet tokens were previously percent encoded. This can cause issues with
clients doing their own encoding or not. By removing the padding and then
re-establishing it when we validate the token, we don't present that problem to
clients. This also shortens the length of a Fernet token.

Change-Id: I674bad86ccc9027ac3b365c10b3b142fc9d73c17
Related-Bug: 1433372
Closes-Bug: 1491926