Updated openstack/openstack Project: openstack/cinder  de64f5ad716676b7180365798efc3ea69a4fef0e
Return a tuple from SolidFire update_provider_info

The SolidFire update_provider_info is only returning
a list of updated volumes, but the manager call (and
the method def in the reference) calls for a tuple of
volumes and snapshots.

This patch fixes the return to be a tuple, a follow on
will be needed to add snapshot updates.

Change-Id: Ifa617bfd7b0652adf7bf1b19d48e88bfcfcec8df
Partial-Bug: #1493563