Updated openstack/openstack Project: openstack/horizon  2aa3cc95286424718d8988a9aa51215c13917bd7
Fix a bug for "Set Gateway" form modal

When set gateway for a router having no name, it cast error:
"This field is required" for the router name. So there
should display the router's name_or_id in "Set Gateway"
form modal, and this is consistent with the "Edit Router".

Change-Id: Ibfcbf90f20c2eeb8b5533f8f41d4efc68df3ffbb
Closes-Bug: #1492667