Updated openstack/openstack Project: openstack/tempest  96e727983ca0ae04be509ac22dfbdc3a363e376e
Switch using token_client to the latest for gate

On tempest-lib 0.8.0, TokenClientJSON is marked as deprecated
and a new TokenClient became standard instead.
In Tempest unit tests, we are stubbing TokenClientJSON but it
doesn't work at all now.
This patch makes these unit tests switch to TokenClient for
passing the gate job.

Change-Id: I6404fd8215a619a90eaeea22afa479639d150872