Updated openstack/openstack Project: openstack/tempest  7c7280edd83c42bb4e8f77fca8cd7dfd03a34a43
Add docs section on config options to reviewers guide

This commit adds an additional section to the reviewers guide to
elaborate on the policy around configuration options. We already
documented part of it in the plugin interface documentation, but this
should make the policy around it a bit more clear.

Change-Id: Ied2d7b3298d613b7e986b31fd0a08ad7c73afa3b