Updated openstack/openstack Project: openstack/horizon  27ceba6035c22b7f11e529ddbc287f27568bb8e8
Fix usage of python-novaclient in Horizon

Direct `from novaclient.v2 import client` became source of problems
since the introduction of microversions in Nova. According to
novaclient docs, `from novaclient import client` (with the following
specification of version 2 in Client constructor) is a proper way of
using it. It should safeguard Horizon against the backward
incompatibility issues, such as in bug 1493205.

Change-Id: Ifedd47e8ad6a7aac52549f6e1742dd73fed48da1
Closes-Bug: #1493576
Closes-Bug: #1493205