Updated openstack/openstack Project: openstack-infra/project-config  ab1e470c272314cb88c2c88510dbeaf2c6b16232
stackforge/networking-bagpipe-l2: add python-jobs

Adds python-jobs for stackforge/networking-bagpipe-l2 project.

The testsuite is currently empty but the tox.ini is clean and
both 'tox -epy17' and 'tox -epy34' run fine.

Note that not having any python-jobs block any merge from hapenning
on the project because the only job is the check-requirement job which
is a conditional job (as was explained to me by fungi).

Hence it would be great to merge this change as quickly as possible.

Change-Id: I3eefd7deefc2319cdc9b6c922394dadae4ae4f52